Remove Popup CSS and Replace With Modal
Popup.css file along with dialog and related codes have been removed as the popup script is not being used and replaced with a Bootstrap Modal. The switch has simplified the process and reduced the complexity of manually generating the styles. The new modal code and associated styles have been added to the profile page. As a part of this change, the codepen link and unnecessary comments in popup.css have been removed. In addition, the version of Bootstrap.js has been upgraded to improve support for modals.